Move the move-only helpers out of test.hpp
A kernel can use them; the identifiers that live alongside them are what nvcc cannot parse.
