+++ test/test.hpp
@@ -9,6 +9,8 @@
 
 #include <sstream>
 
+#include "types.hpp"
+
 using namespace kumi::literals;
 
 namespace kumi
@@ -329,35 +331,3 @@ template<kumi::concepts::record_type R> operations copy_assign_fwd(R&& r)
   local = r["d"_id];
   return local.value;
 }
-
-//==============================================================================================
-//! Type to ensure the correctness of move semantic
-//==============================================================================================
-template<typename T, typename Enable = void> struct is_moveonly_i : std::false_type
-{
-};
-
-template<typename T> struct is_moveonly_i<T, typename T::is_moveonly_i> : std::true_type
-{
-};
-
-template<typename T> struct is_not_moveonly_i : std::bool_constant<!is_moveonly_i<T>::value>
-{
-};
-
-template<typename T> using is_moveonly_type = is_moveonly_i<std::remove_cvref_t<T>>;
-template<typename T> using is_not_moveonly_type = is_not_moveonly_i<std::remove_cvref_t<T>>;
-
-struct moveonly
-{
-  using is_moveonly = void;
-  moveonly() = default;
-  moveonly(moveonly const&) = delete;
-  moveonly(moveonly&&) = default;
-  moveonly& operator=(moveonly const&) = delete;
-  moveonly& operator=(moveonly&&) = default;
-};
-
-struct empty
-{
-};
