+++ test/types.hpp
@@ -1,0 +1,44 @@
+//==================================================================================================
+/*
+  KUMI - Compact Tuple Tools
+  Copyright : KUMI Project Contributors
+  SPDX-License-Identifier: BSL-1.0
+*/
+//==================================================================================================
+#pragma once
+
+#include <type_traits>
+
+// Kept out of test.hpp, whose identifiers nvcc refuses to parse, so that a kernel can use these too.
+
+//==============================================================================================
+//! Type to ensure the correctness of move semantic
+//==============================================================================================
+template<typename T, typename Enable = void> struct is_moveonly_i : std::false_type
+{
+};
+
+template<typename T> struct is_moveonly_i<T, typename T::is_moveonly_i> : std::true_type
+{
+};
+
+template<typename T> struct is_not_moveonly_i : std::bool_constant<!is_moveonly_i<T>::value>
+{
+};
+
+template<typename T> using is_moveonly_type = is_moveonly_i<std::remove_cvref_t<T>>;
+template<typename T> using is_not_moveonly_type = is_not_moveonly_i<std::remove_cvref_t<T>>;
+
+struct moveonly
+{
+  using is_moveonly = void;
+  moveonly() = default;
+  moveonly(moveonly const&) = delete;
+  moveonly(moveonly&&) = default;
+  moveonly& operator=(moveonly const&) = delete;
+  moveonly& operator=(moveonly&&) = default;
+};
+
+struct empty
+{
+};
